test: organize unit tests by source package
