+++ CONTRIBUTING_zh.md
@@ -49,17 +49,8 @@ uv pip install pre-commit==4.6.2
 pre-commit install
 ```
 
-`test` extra 只包含轻量的单元测试依赖。开发新 robot 或修改已有 robot 时，请同时
-安装对应的运行时 extra，例如：
-
-```bash
-uv pip install -e ".[test,libero-pro]"  # LIBERO-Pro
-uv pip install -e ".[test,robocasa]"    # RoboCasa
-uv pip install -e ".[test,robotwin]"    # RoboTwin
-```
-
-只需安装当前改动涉及的 robot extra；模型 checkpoint、模拟器 assets 和其他预置
-资源仍应作为独立输入准备。
+`test` extra 只包含单元测试依赖。贡献新 robot 或为现有 robot 开发功能时，还需安装
+[安装指南](docs/source-zh/rst_source/installation.rst)中对应的 extra。
 
 ### 第 3 步：开发
 
@@ -91,7 +82,7 @@ pre-commit run --all-files
 pytest tests/unit_tests -v
 
 # 开发时运行单个文件中的相关测试。
-pytest tests/unit_tests/cli/test_main_contracts.py -k xxx -x
+pytest tests/unit_tests/rpent/cli/test_main_contracts.py -k xxx -x
 ```
 
 如果检查失败或修改了文件，请检查这些变化并重新运行。
