+++ tests/unit_tests/robots/robotwin/test_robotwin_tool_schema_contracts.py
@@ -1,0 +1,29 @@
+# Copyright 2026 The RPent Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Robot-specific schema contracts for RoboTwin tools."""
+
+from robots.robotwin import tools
+
+
+def test_perception_schemas_use_the_same_view_coordinate_space() -> None:
+    by_name = {spec["name"]: spec for spec in tools.TOOLS_SPEC}
+
+    for tool_name, coordinate_name in (
+        ("sample_world_xyz", "pixels"),
+        ("query_world_map", "bbox"),
+    ):
+        schema = by_name[tool_name]["input_schema"]
+        assert schema["required"] == ["view", coordinate_name]
+        assert schema["properties"]["view"]["type"] == "string"
