+++ tests/unit_tests/rpent/memory/test_cli_contracts.py
