+++ tests/unit_tests/rpent/memory/test_manager_contracts.py
