+++ tests/unit_tests/rpent/planner/test_claude_contracts.py
