+++ tests/unit_tests/rpent/utils/test_daemon_lifecycle.py
